+++ 0x0E-structures_typedef/dog.h
@@ -16,6 +16,6 @@ struct dog
 	char *owner;
 };
 
-int init_dog(struct dog *d, char * name, int age, char * owner);
+int init_dog(struct dog *d, char * name, float age, char * owner);
 
 #endif
